+++ src/commonTest/kotlin/internal/Failures.kt
@@ -1,14 +1,13 @@
 package kotools.types.internal
 
-import kotlin.jvm.JvmSynthetic
 import kotlin.reflect.KClass
+import kotlin.test.fail
 
-@JvmSynthetic
-internal inline fun <reified T : Any> unexpectedCreationError(
+internal inline fun <reified T : Any> unexpectedCreationFailure(
     value: Any
 ): Nothing {
     val kClass: KClass<T> = T::class
     val typeName: String = kClass.simpleName
         ?: error("Getting simple name of '$kClass' shouldn't return 'null'.")
-    error("Creating an instance of $typeName from '$value' shouldn't fail.")
+    fail("Creating a $typeName from '$value' shouldn't fail.")
 }