+++ src/commonTest/kotlin/number/NegativeIntTest.kt
@@ -12,6 +12,7 @@ import kotlinx.serialization.json.Json
 import kotools.types.Package
 import kotools.types.experimental.ExperimentalNumberApi
 import kotools.types.experimental.ExperimentalRangeApi
+import kotools.types.internal.unexpectedCreationFailure
 import kotools.types.range.InclusiveBound
 import kotools.types.range.NotEmptyRange
 import kotools.types.shouldEqual
@@ -76,28 +77,6 @@ class NegativeIntTest {
         val result: Result<NegativeInt> = number.toNegativeInt()
         val exception: IllegalArgumentException = assertFailsWith {
             result.getOrThrow()
-        }
-        val actualMessage: String = assertNotNull(exception.message)
-        val expectedMessage: String = NegativeIntConstructionException(number)
-            .message
-        assertEquals(expectedMessage, actualMessage)
-    }
-
-    @ExperimentalNumberApi
-    @Test
-    fun toNegativeIntOrThrow_should_pass_with_a_negative_Int() {
-        val expected: Number = Random.nextInt(from = Int.MIN_VALUE, until = 0)
-        val result: NegativeInt = expected.toNegativeIntOrThrow()
-        val actual: Int = result.toInt()
-        assertEquals(expected, actual)
-    }
-
-    @ExperimentalNumberApi
-    @Test
-    fun toNegativeIntOrThrow_should_fail_with_a_strictly_positive_Int() {
-        val number: Number = Random.nextInt(from = 1, until = Int.MAX_VALUE)
-        val exception: IllegalArgumentException = assertFailsWith {
-            number.toNegativeIntOrThrow()
         }
         val actualMessage: String = assertNotNull(exception.message)
         val expectedMessage: String = NegativeIntConstructionException(number)
@@ -178,3 +157,7 @@ class NegativeIntSerializerTest {
         exception.shouldHaveAMessage()
     }
 }
+
+internal fun Number.toNegativeIntOrFailure(): NegativeInt = toNegativeInt()
+    .getOrNull()
+    ?: unexpectedCreationFailure<NegativeInt>(value = this)