+++ CHANGELOG.md
@@ -70,7 +70,7 @@ All notable changes to this project will be documented in this file.
 - The following **experimental** factory functions (issue [#258]):
   - `toNonZeroIntOrNull` and `toNonZeroIntOrThrow`
   - `toPositiveIntOrNull` and `toPositiveIntOrThrow`
-  - `toNegativeIntOrNull`.
+  - `toNegativeIntOrNull` and `toNegativeIntOrThrow`.
 
 [#258]: https://github.com/kotools/types/issues/258